requirements.txt
gunicorn
flask
pymysql
peewee
redis
qiniu
requests
pycrypto
ipgetter
xmltodict
celery>=4.0,<5.0
